pyproject.toml
[tool.poetry]
name = "config2class"
version = "0.1.0"
description = ""
authors = ["Robin Uhrich <robin.uhrich@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pyaml = "^24.0.0"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
